style: Renamed tests
