+++ tests/PowerUtils.Results.Tests/ValueResults/FactorySuccessTests.cs
@@ -4,10 +4,10 @@ using Xunit;
 
 namespace PowerUtils.Results.Tests.ValueResults
 {
-    public class MethodSuccessTests
+    public class FactorySuccessTests
     {
         [Fact]
-        public void SuccessMethod_CreateResult_IsErrorFalse()
+        public void SuccessFactory_CreateResult_IsErrorFalse()
         {
             // Arrange && Act
             var act = Result.Success(new FakeModel());
